Added git message encode for CAN bus
It works a little different from the stream based approaches for
CBOR and JSON as we have to produce CAN messages.
The new method places the relevant message data for a single object
in externally provided buffers. This makes it possible to have no hardware
related code in.

A new helper to get the channel by id was added, too.